fix: cargo fmt + clippy fixes across crates
- Fix formatting across all crates
- Add Default impl for HotFrameCache
- Remove nested format! in archive.rs
- Fix needless borrow in macos accessibility

+++ crates/screenpipe-server/src/event_driven_capture.rs
@@ -270,8 +270,7 @@ pub async fn event_driven_capture_loop(
                     last_content_hash = result.content_hash;
                     last_db_write = Instant::now();
                     vision_metrics.record_capture();
-                    vision_metrics
-                        .record_db_write(Duration::from_millis(result.duration_ms as u64));
+                    vision_metrics.record_db_write(Duration::from_millis(result.duration_ms));
                     if let Some(ref cache) = hot_frame_cache {
                         push_to_hot_cache(cache, result, &device_name, &CaptureTrigger::Manual)
                             .await;
@@ -408,7 +407,7 @@ pub async fn event_driven_capture_loop(
                             last_db_write = Instant::now();
                             vision_metrics.record_capture();
                             vision_metrics
-                                .record_db_write(Duration::from_millis(result.duration_ms as u64));
+                                .record_db_write(Duration::from_millis(result.duration_ms));
 
                             if let Some(ref cache) = hot_frame_cache {
                                 push_to_hot_cache(cache, result, &device_name, &trigger).await;
